check permissions for deletion

+++ app/controllers/jobs_controller.rb
@@ -78,39 +78,51 @@ class JobsController < ApplicationController
   end
 
   def destroy
-    thisjob = Job.find_by(:job_id => params[:id])
-    if thisjob then
-      flash[:info] = "Job '#{thisjob[:name]}' was deleted."
-      thisjob.destroy
-      queue = Sidekiq::Queue.new
-      queue.each do |job|
-        job.delete if job.jid == params[:id]
-      end
-      if File.exist?("#{thisjob.job_directory}") then
-        FileUtils.rm_rf("#{thisjob.job_directory}")
-      end
-    end
-    if logged_in? then
-      redirect_to :jobs
+    if !logged_in? then
+      flash[:info] = "You do not have permission to delete jobs in the " + \
+                     "queue."
+      redirect_to :welcome
     else
-      redirect_to "welcome/index"
+      thisjob = Job.find_by(:job_id => params[:id])
+      if thisjob then
+        flash[:info] = "Job '#{thisjob[:name]}' was deleted."
+        thisjob.destroy
+        queue = Sidekiq::Queue.new
+        queue.each do |job|
+          job.delete if job.jid == params[:id]
+        end
+        if File.exist?("#{thisjob.job_directory}") then
+          FileUtils.rm_rf("#{thisjob.job_directory}")
+        end
+      end
+      if logged_in? then
+        redirect_to :jobs
+      else
+        redirect_to "welcome/index"
+      end
     end
   end
 
   def destroy_by_int_id
-    thisjob = Job.find_by(:job_id => params[:id])
-    if thisjob then
-      flash.now[:info] = "Deleted job #{thisjob[:name]}"
-      thisjob.destroy
-      queue = Sidekiq::Queue.new
-      queue.each do |job|
-        job.delete if job.jid == params[:id]
+    if !logged_in? then
+      flash[:info] = "You do not have permission to delete jobs in the " + \
+                     "queue."
+      redirect_to :welcome
+    else
+      thisjob = Job.find_by(:job_id => params[:id])
+      if thisjob then
+        flash.now[:info] = "Deleted job #{thisjob[:name]}"
+        thisjob.destroy
+        queue = Sidekiq::Queue.new
+        queue.each do |job|
+          job.delete if job.jid == params[:id]
+        end
+        if File.exist?("#{thisjob.job_directory}") then
+          FileUtils.rm_rf("#{thisjob.job_directory}")
+        end
       end
-      if File.exist?("#{thisjob.job_directory}") then
-        FileUtils.rm_rf("#{thisjob.job_directory}")
-      end
+      render "welcome/index"
     end
-    render "welcome/index"
   end
 
   def show
